+++ lnbits/core/extensions/models.py
@@ -1,6 +1,5 @@
 import asyncio
 import hashlib
-import importlib
 import json
 import os
 import shutil
@@ -8,15 +7,20 @@ import sys
 import zipfile
 from pathlib import Path
 from typing import Any, List, NamedTuple, Optional, Tuple
-from urllib import request
 
 import httpx
 from loguru import logger
-from packaging import version
 from pydantic import BaseModel
 
-from lnbits.core.crud import delete_installed_extension, get_installed_extension
 from lnbits.settings import settings
+
+from .helpers import (
+    download_url,
+    file_hash,
+    github_api_get,
+    icon_to_github_url,
+    version_parse,
+)
 
 
 class ExplicitRelease(BaseModel):
@@ -124,124 +128,6 @@ class UserExtension(BaseModel):
         return ext
 
 
-def download_url(url, save_path):
-    with request.urlopen(url, timeout=60) as dl_file:
-        with open(save_path, "wb") as out_file:
-            out_file.write(dl_file.read())
-
-
-def file_hash(filename):
-    h = hashlib.sha256()
-    b = bytearray(128 * 1024)
-    mv = memoryview(b)
-    with open(filename, "rb", buffering=0) as f:
-        while n := f.readinto(mv):
-            h.update(mv[:n])
-    return h.hexdigest()
-
-
-async def fetch_github_repo_info(
-    org: str, repository: str
-) -> Tuple[GitHubRepo, GitHubRepoRelease, ExtensionConfig]:
-    repo_url = f"https://api.github.com/repos/{org}/{repository}"
-    error_msg = "Cannot fetch extension repo"
-    repo = await github_api_get(repo_url, error_msg)
-    github_repo = GitHubRepo.parse_obj(repo)
-
-    lates_release_url = (
-        f"https://api.github.com/repos/{org}/{repository}/releases/latest"
-    )
-    error_msg = "Cannot fetch extension releases"
-    latest_release: Any = await github_api_get(lates_release_url, error_msg)
-
-    config_url = f"https://raw.githubusercontent.com/{org}/{repository}/{github_repo.default_branch}/config.json"
-    error_msg = "Cannot fetch config for extension"
-    config = await github_api_get(config_url, error_msg)
-
-    return (
-        github_repo,
-        GitHubRepoRelease.parse_obj(latest_release),
-        ExtensionConfig.parse_obj(config),
-    )
-
-
-async def fetch_manifest(url) -> Manifest:
-    error_msg = "Cannot fetch extensions manifest"
-    manifest = await github_api_get(url, error_msg)
-    return Manifest.parse_obj(manifest)
-
-
-async def fetch_github_releases(org: str, repo: str) -> List[GitHubRepoRelease]:
-    releases_url = f"https://api.github.com/repos/{org}/{repo}/releases"
-    error_msg = "Cannot fetch extension releases"
-    releases = await github_api_get(releases_url, error_msg)
-    return [GitHubRepoRelease.parse_obj(r) for r in releases]
-
-
-async def fetch_github_release_config(
-    org: str, repo: str, tag_name: str
-) -> Optional[ExtensionConfig]:
-    config_url = (
-        f"https://raw.githubusercontent.com/{org}/{repo}/{tag_name}/config.json"
-    )
-    error_msg = "Cannot fetch GitHub extension config"
-    config = await github_api_get(config_url, error_msg)
-    return ExtensionConfig.parse_obj(config)
-
-
-async def github_api_get(url: str, error_msg: Optional[str]) -> Any:
-    headers = {"User-Agent": settings.user_agent}
-    if settings.lnbits_ext_github_token:
-        headers["Authorization"] = f"Bearer {settings.lnbits_ext_github_token}"
-    async with httpx.AsyncClient(headers=headers) as client:
-        resp = await client.get(url)
-        if resp.status_code != 200:
-            logger.warning(f"{error_msg} ({url}): {resp.text}")
-        resp.raise_for_status()
-        return resp.json()
-
-
-async def fetch_release_payment_info(
-    url: str, amount: Optional[int] = None
-) -> Optional[ReleasePaymentInfo]:
-    if amount:
-        url = f"{url}?amount={amount}"
-    try:
-        async with httpx.AsyncClient() as client:
-            resp = await client.get(url)
-            resp.raise_for_status()
-            return ReleasePaymentInfo(**resp.json())
-    except Exception as e:
-        logger.warning(e)
-        return None
-
-
-async def fetch_release_details(details_link: str) -> Optional[dict]:
-
-    try:
-        async with httpx.AsyncClient() as client:
-            resp = await client.get(details_link)
-            resp.raise_for_status()
-            data = resp.json()
-            if "description_md" in data:
-                resp = await client.get(data["description_md"])
-                if not resp.is_error:
-                    data["description_md"] = resp.text
-
-            return data
-    except Exception as e:
-        logger.warning(e)
-        return None
-
-
-def icon_to_github_url(source_repo: str, path: Optional[str]) -> str:
-    if not path:
-        return ""
-    _, _, *rest = path.split("/")
-    tail = "/".join(rest)
-    return f"https://github.com/{source_repo}/raw/main/{tail}"
-
-
 class Extension(NamedTuple):
     code: str
     is_valid: bool
@@ -281,50 +167,6 @@ class Extension(NamedTuple):
         )
 
 
-# All subdirectories in the current directory, not recursive.
-
-
-class ExtensionManager:
-    def __init__(self) -> None:
-        p = Path(settings.lnbits_extensions_path, "extensions")
-        Path(p).mkdir(parents=True, exist_ok=True)
-        self._extension_folders: List[Path] = [f for f in p.iterdir() if f.is_dir()]
-
-    @property
-    def extensions(self) -> List[Extension]:
-        # todo: remove this property somehow, it is too expensive
-        output: List[Extension] = []
-
-        for extension_folder in self._extension_folders:
-            extension_code = extension_folder.parts[-1]
-            try:
-                with open(extension_folder / "config.json") as json_file:
-                    config = json.load(json_file)
-                is_valid = True
-                is_admin_only = extension_code in settings.lnbits_admin_extensions
-            except Exception:
-                config = {}
-                is_valid = False
-                is_admin_only = False
-
-            output.append(
-                Extension(
-                    extension_code,
-                    is_valid,
-                    is_admin_only,
-                    config.get("name"),
-                    config.get("short_description"),
-                    config.get("tile"),
-                    config.get("contributors"),
-                    config.get("hidden") or False,
-                    config.get("migration_module"),
-                    config.get("db_name"),
-                )
-            )
-
-        return output
-
-
 class ExtensionRelease(BaseModel):
     name: str
     version: str
@@ -358,8 +200,22 @@ class ExtensionRelease(BaseModel):
         if not self.pay_link:
             return
 
-        payment_info = await fetch_release_payment_info(self.pay_link)
+        payment_info = await self.fetch_release_payment_info()
         self.cost_sats = payment_info.amount if payment_info else None
+
+    async def fetch_release_payment_info(
+        self, amount: Optional[int] = None
+    ) -> Optional[ReleasePaymentInfo]:
+        url = f"{self.pay_link}?amount={amount}" if amount else self.pay_link
+        assert url, "Missing URL for payment info."
+        try:
+            async with httpx.AsyncClient() as client:
+                resp = await client.get(url)
+                resp.raise_for_status()
+                return ReleasePaymentInfo(**resp.json())
+        except Exception as e:
+            logger.warning(e)
+            return None
 
     @classmethod
     def from_github_release(
@@ -401,7 +257,7 @@ class ExtensionRelease(BaseModel):
     @classmethod
     async def get_github_releases(cls, org: str, repo: str) -> List["ExtensionRelease"]:
         try:
-            github_releases = await fetch_github_releases(org, repo)
+            github_releases = await cls.fetch_github_releases(org, repo)
             return [
                 ExtensionRelease.from_github_release(f"{org}/{repo}", r)
                 for r in github_releases
@@ -409,6 +265,15 @@ class ExtensionRelease(BaseModel):
         except Exception as e:
             logger.warning(e)
             return []
+
+    @classmethod
+    async def fetch_github_releases(
+        cls, org: str, repo: str
+    ) -> List[GitHubRepoRelease]:
+        releases_url = f"https://api.github.com/repos/{org}/{repo}/releases"
+        error_msg = "Cannot fetch extension releases"
+        releases = await github_api_get(releases_url, error_msg)
+        return [GitHubRepoRelease.parse_obj(r) for r in releases]
 
 
 class InstallableExtension(BaseModel):
@@ -638,7 +503,7 @@ class InstallableExtension(BaseModel):
         cls, github_release: GitHubRelease
     ) -> Optional["InstallableExtension"]:
         try:
-            repo, latest_release, config = await fetch_github_repo_info(
+            repo, latest_release, config = await cls.fetch_github_repo_info(
                 github_release.organisation, github_release.repository
             )
             source_repo = f"{github_release.organisation}/{github_release.repository}"
@@ -679,7 +544,7 @@ class InstallableExtension(BaseModel):
 
         for url in settings.lnbits_extensions_manifests:
             try:
-                manifest = await fetch_manifest(url)
+                manifest = await cls.fetch_manifest(url)
 
                 for r in manifest.repos:
                     ext = await InstallableExtension.from_github_release(r)
@@ -720,7 +585,7 @@ class InstallableExtension(BaseModel):
 
         for url in settings.lnbits_extensions_manifests:
             try:
-                manifest = await fetch_manifest(url)
+                manifest = await cls.fetch_manifest(url)
                 for r in manifest.repos:
                     if r.id != ext_id:
                         continue
@@ -758,6 +623,37 @@ class InstallableExtension(BaseModel):
 
         return selected_release[0] if len(selected_release) != 0 else None
 
+    @classmethod
+    async def fetch_github_repo_info(
+        cls, org: str, repository: str
+    ) -> Tuple[GitHubRepo, GitHubRepoRelease, ExtensionConfig]:
+        repo_url = f"https://api.github.com/repos/{org}/{repository}"
+        error_msg = "Cannot fetch extension repo"
+        repo = await github_api_get(repo_url, error_msg)
+        github_repo = GitHubRepo.parse_obj(repo)
+
+        lates_release_url = (
+            f"https://api.github.com/repos/{org}/{repository}/releases/latest"
+        )
+        error_msg = "Cannot fetch extension releases"
+        latest_release: Any = await github_api_get(lates_release_url, error_msg)
+
+        config_url = f"https://raw.githubusercontent.com/{org}/{repository}/{github_repo.default_branch}/config.json"
+        error_msg = "Cannot fetch config for extension"
+        config = await github_api_get(config_url, error_msg)
+
+        return (
+            github_repo,
+            GitHubRepoRelease.parse_obj(latest_release),
+            ExtensionConfig.parse_obj(config),
+        )
+
+    @classmethod
+    async def fetch_manifest(cls, url) -> Manifest:
+        error_msg = "Cannot fetch extensions manifest"
+        manifest = await github_api_get(url, error_msg)
+        return Manifest.parse_obj(manifest)
+
 
 class CreateExtension(BaseModel):
     ext_id: str
@@ -772,74 +668,3 @@ class ExtensionDetailsRequest(BaseModel):
     ext_id: str
     source_repo: str
     version: str
-
-
-async def uninstall_extension(ext_id):
-    await stop_extension_background_work(ext_id)
-
-    settings.lnbits_deactivated_extensions.add(ext_id)
-
-    extension = await get_installed_extension(ext_id)
-    if extension:
-        extension.clean_extension_files()
-    await delete_installed_extension(ext_id=ext_id)
-
-
-async def stop_extension_background_work(ext_id) -> bool:
-    """
-    Stop background work for extension (like asyncio.Tasks, WebSockets, etc).
-    Extensions SHOULD expose a `api_stop()` function.
-    """
-    upgrade_hash = settings.extension_upgrade_hash(ext_id) or ""
-    ext = Extension(ext_id, True, False, upgrade_hash=upgrade_hash)
-
-    try:
-        logger.info(f"Stopping background work for extension '{ext.module_name}'.")
-        old_module = importlib.import_module(ext.module_name)
-
-        # Extensions must expose an `{ext_id}_stop()` function at the module level
-        # The `api_stop()` function is for backwards compatibility (will be deprecated)
-        stop_fns = [f"{ext_id}_stop", "api_stop"]
-        stop_fn_name = next((fn for fn in stop_fns if hasattr(old_module, fn)), None)
-        assert stop_fn_name, "No stop function found for '{ext.module_name}'"
-
-        stop_fn = getattr(old_module, stop_fn_name)
-        if stop_fn:
-            await stop_fn()
-
-        logger.info(f"Stopped background work for extension '{ext.module_name}'.")
-    except Exception as ex:
-        logger.warning(f"Failed to stop background work for '{ext.module_name}'.")
-        logger.warning(ex)
-        return False
-
-    return True
-
-
-def get_valid_extensions(include_deactivated: Optional[bool] = True) -> List[Extension]:
-    valid_extensions = [
-        extension for extension in ExtensionManager().extensions if extension.is_valid
-    ]
-
-    if include_deactivated:
-        return valid_extensions
-
-    if settings.lnbits_extensions_deactivate_all:
-        return []
-
-    return [
-        e
-        for e in valid_extensions
-        if e.code not in settings.lnbits_deactivated_extensions
-    ]
-
-
-def version_parse(v: str):
-    """
-    Wrapper for version.parse() that does not throw if the version is invalid.
-    Instead it return the lowest possible version ("0.0.0")
-    """
-    try:
-        return version.parse(v)
-    except Exception:
-        return version.parse("0.0.0")
